fix: support List kinds in local lookup resources
