+++ pkg/action/chart_lint.go
@@ -159,11 +159,6 @@ func ChartLint(ctx context.Context, opts ChartLintOptions) error {
 		return fmt.Errorf("local lookup resources are not allowed together with remote mode")
 	}
 
-	localLookupResources, err := parseLocalLookupResources(opts.LocalLookupResourcesPaths)
-	if err != nil {
-		return fmt.Errorf("parse local lookup resources: %w", err)
-	}
-
 	if !opts.Remote {
 		opts.ReleaseStorageDriver = common.ReleaseStorageDriverMemory
 	}
@@ -279,7 +274,7 @@ func ChartLint(ctx context.Context, opts ChartLintOptions) error {
 		ExtraAPIVersions:           opts.ExtraAPIVersions,
 		HelmOptions:                helmOptions,
 		LocalKubeVersion:           opts.LocalKubeVersion,
-		LocalLookupResources:       localLookupResources,
+		LocalLookupResourcesPaths:  opts.LocalLookupResourcesPaths,
 		Remote:                     opts.Remote,
 		TemplatesAllowDNS:          opts.TemplatesAllowDNS,
 		TempDirPath:                opts.TempDirPath,
