+++ pkg/action/common.go
@@ -15,9 +15,6 @@ import (
 	"github.com/gookit/color"
 	"github.com/samber/lo"
 	"github.com/xo/terminfo"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/client-go/kubernetes/scheme"
 
 	"github.com/werf/kubedog/pkg/informer"
 	"github.com/werf/kubedog/pkg/trackers/dyntracker/logstore"
@@ -25,12 +22,10 @@ import (
 	kdutil "github.com/werf/kubedog/pkg/trackers/dyntracker/util"
 	"github.com/werf/nelm/pkg/common"
 	helmrelease "github.com/werf/nelm/pkg/helm/pkg/release"
-	"github.com/werf/nelm/pkg/helm/pkg/releaseutil"
 	"github.com/werf/nelm/pkg/kube"
 	"github.com/werf/nelm/pkg/log"
 	"github.com/werf/nelm/pkg/plan"
 	"github.com/werf/nelm/pkg/release"
-	"github.com/werf/nelm/pkg/resource/spec"
 	"github.com/werf/nelm/pkg/util"
 )
 
@@ -118,74 +113,6 @@ func handleBuildPlanErr(ctx context.Context, installPlan *plan.Plan, planErr err
 	}
 
 	log.Default.Warn(ctx, "Plan graph saved to %q for debugging", graphPath)
-}
-
-func parseLocalLookupResources(paths []string) ([]*unstructured.Unstructured, error) {
-	var resources []*unstructured.Unstructured
-
-	seen := make(map[string]bool)
-
-	for _, path := range paths {
-		content, err := os.ReadFile(path)
-		if err != nil {
-			return nil, fmt.Errorf("read file %q: %w", path, err)
-		}
-
-		for i, manifest := range releaseutil.SplitManifestsToSlice(string(content)) {
-			obj, _, err := scheme.Codecs.UniversalDecoder().Decode([]byte(manifest), nil, &unstructured.Unstructured{})
-			if err != nil {
-				return nil, fmt.Errorf("parse file %q (document %d): %w", path, i, err)
-			}
-
-			unstruct := obj.(*unstructured.Unstructured)
-
-			if unstruct.IsList() {
-				item := 0
-
-				if err := unstruct.EachListItem(func(o runtime.Object) error {
-					res, err := collectLocalLookupResource(o.(*unstructured.Unstructured), seen)
-					if err != nil {
-						return err
-					}
-
-					item++
-					resources = append(resources, res)
-
-					return nil
-				}); err != nil {
-					return nil, fmt.Errorf("parse file %q (document %d, item %d): %w", path, i, item, err)
-				}
-
-				continue
-			}
-
-			res, err := collectLocalLookupResource(unstruct, seen)
-			if err != nil {
-				return nil, fmt.Errorf("parse file %q (document %d): %w", path, i, err)
-			}
-
-			resources = append(resources, res)
-		}
-	}
-
-	return resources, nil
-}
-
-func collectLocalLookupResource(unstruct *unstructured.Unstructured, seen map[string]bool) (*unstructured.Unstructured, error) {
-	if unstruct.GetAPIVersion() == "" {
-		return nil, fmt.Errorf("apiVersion is missing")
-	}
-
-	gvk := unstruct.GroupVersionKind()
-	id := spec.IDWithVersion(unstruct.GetName(), unstruct.GetNamespace(), gvk.Group, gvk.Version, gvk.Kind)
-
-	if seen[id] {
-		return nil, fmt.Errorf("duplicate resource %s", spec.IDHuman(unstruct.GetName(), unstruct.GetNamespace(), gvk.Group, gvk.Kind))
-	}
-
-	seen[id] = true
-
-	return unstruct, nil
 }
 
 func printNotes(ctx context.Context, notes string) {
