+++ pkg/chart/chart_render.go
@@ -15,7 +15,9 @@ import (
 	"github.com/goccy/go-yaml"
 	"github.com/samber/lo"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
+	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/client-go/discovery"
+	"k8s.io/client-go/kubernetes/scheme"
 
 	"github.com/werf/nelm/pkg/common"
 	"github.com/werf/nelm/pkg/featgate"
@@ -44,21 +46,21 @@ type RenderChartOptions struct {
 	common.ChartRepoConnectionOptions
 	common.ValuesOptions
 
-	ChartProvenanceKeyring  string
-	ChartProvenanceStrategy string
-	ChartRepoNoUpdate       bool
-	ChartVersion            string
-	DenoBinaryPath          string
-	ExtraAPIVersions        []string
-	HelmOptions             helmopts.HelmOptions
-	IgnoreBundleJS          bool
-	LocalKubeVersion        string
-	LocalLookupResources    []*unstructured.Unstructured
-	NoStandaloneCRDs        bool
-	Remote                  bool
-	SubchartNotes           bool
-	TempDirPath             string
-	TemplatesAllowDNS       bool
+	ChartProvenanceKeyring    string
+	ChartProvenanceStrategy   string
+	ChartRepoNoUpdate         bool
+	ChartVersion              string
+	DenoBinaryPath            string
+	ExtraAPIVersions          []string
+	HelmOptions               helmopts.HelmOptions
+	IgnoreBundleJS            bool
+	LocalKubeVersion          string
+	LocalLookupResourcesPaths []string
+	NoStandaloneCRDs          bool
+	Remote                    bool
+	SubchartNotes             bool
+	TempDirPath               string
+	TemplatesAllowDNS         bool
 }
 
 type RenderChartResult struct {
@@ -199,8 +201,13 @@ func RenderChart(ctx context.Context, chartPath, releaseName, releaseNamespace s
 		engine = lo.ToPtr(helmengine.New(clientFactory.KubeConfig().RestConfig))
 	} else {
 		engine = &helmengine.Engine{}
-		if len(opts.LocalLookupResources) > 0 {
-			engine.SetClientProvider(lookup.NewLocalClientProvider(opts.LocalLookupResources))
+		if len(opts.LocalLookupResourcesPaths) > 0 {
+			localLookupResources, err := parseLocalLookupResources(opts.LocalLookupResourcesPaths)
+			if err != nil {
+				return nil, fmt.Errorf("parse local lookup resources: %w", err)
+			}
+
+			engine.SetClientProvider(lookup.NewLocalClientProvider(localLookupResources))
 		}
 	}
 
@@ -432,4 +439,72 @@ func validateChart(ctx context.Context, chart *helmchart.Chart) error {
 	}
 
 	return nil
+}
+
+func parseLocalLookupResources(paths []string) ([]*unstructured.Unstructured, error) {
+	var resources []*unstructured.Unstructured
+
+	seen := make(map[string]bool)
+
+	for _, path := range paths {
+		content, err := os.ReadFile(path)
+		if err != nil {
+			return nil, fmt.Errorf("read file %q: %w", path, err)
+		}
+
+		for i, manifest := range releaseutil.SplitManifestsToSlice(string(content)) {
+			obj, _, err := scheme.Codecs.UniversalDecoder().Decode([]byte(manifest), nil, &unstructured.Unstructured{})
+			if err != nil {
+				return nil, fmt.Errorf("parse file %q (document %d): %w", path, i, err)
+			}
+
+			unstruct := obj.(*unstructured.Unstructured)
+
+			if unstruct.IsList() {
+				item := 0
+
+				if err := unstruct.EachListItem(func(o runtime.Object) error {
+					res, err := collectLocalLookupResource(o.(*unstructured.Unstructured), seen)
+					if err != nil {
+						return err
+					}
+
+					item++
+					resources = append(resources, res)
+
+					return nil
+				}); err != nil {
+					return nil, fmt.Errorf("parse file %q (document %d, item %d): %w", path, i, item, err)
+				}
+
+				continue
+			}
+
+			res, err := collectLocalLookupResource(unstruct, seen)
+			if err != nil {
+				return nil, fmt.Errorf("parse file %q (document %d): %w", path, i, err)
+			}
+
+			resources = append(resources, res)
+		}
+	}
+
+	return resources, nil
+}
+
+func collectLocalLookupResource(unstruct *unstructured.Unstructured, seen map[string]bool) (*unstructured.Unstructured, error) {
+	if unstruct.GetAPIVersion() == "" {
+		return nil, fmt.Errorf("apiVersion is missing")
+	}
+
+	gvk := unstruct.GroupVersionKind()
+	id := spec.IDWithVersion(unstruct.GetName(), unstruct.GetNamespace(), gvk.Group, gvk.Version, gvk.Kind)
+
+	if seen[id] {
+		return nil, fmt.Errorf("duplicate resource %s", spec.IDHuman(unstruct.GetName(), unstruct.GetNamespace(), gvk.Group, gvk.Kind))
+	}
+
+	seen[id] = true
+
+	return unstruct, nil
 }
