+++ pkg/chart/parse_local_lookup_resources_ai_test.go
@@ -1,6 +1,6 @@
 //go:build ai_tests
 
-package action
+package chart
 
 import (
 	"os"
